Fix travis CI error

+++ src/core/thread/meshcop_dataset_manager.cpp
@@ -375,9 +375,9 @@ ThreadError DatasetManager::SendSetRequest(otOperationalDataset &aDataset, const
     if (aDataset.mIsNetworkNameSet)
     {
         tlv.SetType(OT_MESHCOP_TLV_NETWORKNAME);
-        tlv.SetLength(strlen(aDataset.mNetworkName.m8));
+        tlv.SetLength(static_cast<uint8_t>(strlen(aDataset.mNetworkName.m8)));
         SuccessOrExit(error = message->Append(&tlv, sizeof(tlv)));
-        SuccessOrExit(error = message->Append(aDataset.mNetworkName.m8, strlen(aDataset.mNetworkName.m8)));
+        SuccessOrExit(error = message->Append(aDataset.mNetworkName.m8, static_cast<uint8_t>(strlen(aDataset.mNetworkName.m8))));
     }
 
     if (aDataset.mIsExtendedPanIdSet)
